Make the timedemo related timers double.
Their being float didn't cause noticeable trouble when time was 0-based,
but with it 4Gs based, timedemo results got very weird.

+++ nq/include/client.h
@@ -116,7 +116,7 @@ typedef struct {
 	qboolean    timedemo;
 	int         td_lastframe;		// to meter out one message a frame
 	int         td_startframe;		// host_framecount at start
-	float       td_starttime;		// realtime at second frame of timedemo
+	double      td_starttime;		// realtime at second frame of timedemo
 } client_static_t;
 
 extern client_static_t	cls;
